Main GUI classes. Settings framework.
Applications settings framework and API
- Workspace configuration files using new settings API (continued)
- Compatibility with BRAT 3.3.0 and with whole application usage.

Main GUI classes
- Display editors core functionality and sub-window management.
- Actions Table
- Collapsible widget groups.
- Menu and toolbars actions
- More resources.
- Ports to all systems

+++ new-gui/brat-lab/QbrtMainWindow.cpp
@@ -247,9 +247,9 @@ void QbrtMainWindow::checkIfStuffDone()
 
 		QString message;
 		if ( mDone == mTotal )
-			message = tr( "All %n image(s) converted", "", mDone );
+			message = tr( "All %n image(s) converted", "", (int)mDone );
 		else
-			message = tr( "Converted %n/%1 image(s)", "", mDone )
+			message = tr( "Converted %n/%1 image(s)", "", (int)mDone )
 			.arg( mTotal );
 
 		mMessageButton->setText( message );
